convert lightsource forms to use custom hook. update logic for show/hide pagination & pagination title. remove collections data from Collection component, and move functionality for add collection/room, filter collection/room up into the dashboard component.
